+++ src/algorithms/local_search/operator.h
@@ -5,14 +5,13 @@
 
 This file is part of VROOM.
 
-Copyright (c) 2015-2020, Julien Coupey.
+Copyright (c) 2015-2021, Julien Coupey.
 All rights reserved (see LICENSE).
 
 */
 
 #include "structures/typedefs.h"
 #include "structures/vroom/input/input.h"
-#include "structures/vroom/raw_route.h"
 #include "structures/vroom/solution_state.h"
 
 namespace vroom {