+++ src/algorithms/munkres.h
@@ -5,7 +5,7 @@
 
 This file is part of VROOM.
 
-Copyright (c) 2015-2020, Julien Coupey.
+Copyright (c) 2015-2021, Julien Coupey.
 All rights reserved (see LICENSE).
 
 */
@@ -13,7 +13,6 @@ All rights reserved (see LICENSE).
 #include <unordered_map>
 
 #include "structures/generic/matrix.h"
-#include "structures/typedefs.h"
 
 namespace vroom {
 namespace utils {