+++ src/problems/tsp/heuristics/christofides.cpp
@@ -2,13 +2,12 @@
 
 This file is part of VROOM.
 
-Copyright (c) 2015-2020, Julien Coupey.
+Copyright (c) 2015-2021, Julien Coupey.
 All rights reserved (see LICENSE).
 
 */
 
 #include <cassert>
-#include <random>
 #include <set>
 
 #include "algorithms/kruskal.h"