+++ src/problems/vrptw/operators/reverse_two_opt.cpp
@@ -2,12 +2,10 @@
 
 This file is part of VROOM.
 
-Copyright (c) 2015-2020, Julien Coupey.
+Copyright (c) 2015-2021, Julien Coupey.
 All rights reserved (see LICENSE).
 
 */
-
-#include <iostream>
 
 #include "problems/vrptw/operators/reverse_two_opt.h"