+++ src/structures/generic/undirected_graph.h
@@ -5,14 +5,12 @@
 
 This file is part of VROOM.
 
-Copyright (c) 2015-2020, Julien Coupey.
+Copyright (c) 2015-2021, Julien Coupey.
 All rights reserved (see LICENSE).
 
 */
 
-#include <list>
 #include <unordered_map>
-#include <vector>
 
 #include "structures/generic/edge.h"
 #include "structures/generic/matrix.h"