+++ src/structures/vroom/input/input.cpp
@@ -2,19 +2,15 @@
 
 This file is part of VROOM.
 
-Copyright (c) 2015-2020, Julien Coupey.
+Copyright (c) 2015-2021, Julien Coupey.
 All rights reserved (see LICENSE).
 
 */
-#include <array>
 
+#include "structures/vroom/input/input.h"
 #include "algorithms/validation/check.h"
 #include "problems/cvrp/cvrp.h"
-#include "problems/tsp/tsp.h"
-#include "problems/vrp.h"
 #include "problems/vrptw/vrptw.h"
-#include "structures/vroom/input/input.h"
-#include "utils/exception.h"
 #include "utils/helpers.h"
 
 namespace vroom {