+++ src/structures/vroom/solution_state.h
@@ -5,17 +5,13 @@
 
 This file is part of VROOM.
 
-Copyright (c) 2015-2020, Julien Coupey.
+Copyright (c) 2015-2021, Julien Coupey.
 All rights reserved (see LICENSE).
 
 */
 
-#include <unordered_set>
-
 #include "structures/typedefs.h"
-#include "structures/vroom/amount.h"
 #include "structures/vroom/input/input.h"
-#include "structures/vroom/raw_route.h"
 #include "structures/vroom/tw_route.h"
 
 namespace vroom {