+++ src/structures/vroom/tw_route.h
@@ -5,13 +5,10 @@
 
 This file is part of VROOM.
 
-Copyright (c) 2015-2020, Julien Coupey.
+Copyright (c) 2015-2021, Julien Coupey.
 All rights reserved (see LICENSE).
 
 */
-
-#include <numeric>
-#include <vector>
 
 #include "structures/typedefs.h"
 #include "structures/vroom/input/input.h"