+++ src/utils/output_json.h
@@ -5,12 +5,10 @@
 
 This file is part of VROOM.
 
-Copyright (c) 2015-2020, Julien Coupey.
+Copyright (c) 2015-2021, Julien Coupey.
 All rights reserved (see LICENSE).
 
 */
-
-#include <string>
 
 #include "../include/rapidjson/document.h"
 #include "structures/vroom/solution/solution.h"